+++ src/rgb/contract/seal.rs
@@ -28,7 +28,7 @@ pub enum Revealed {
     /// Seal that is revealed
     TxOutpoint(OutpointReveal),
     /// Seal contained within the witness transaction
-    WitnessVout { vout: u16, blinding: u32 },
+    WitnessVout { vout: u16, blinding: u64 },
 }
 
 impl Revealed {
@@ -105,7 +105,7 @@ mod strict_encoding {
                 0u8 => Revealed::TxOutpoint(OutpointReveal::strict_decode(d)?),
                 1u8 => Revealed::WitnessVout {
                     vout: u16::strict_decode(&mut d)?,
-                    blinding: u32::strict_decode(&mut d)?,
+                    blinding: u64::strict_decode(&mut d)?,
                 },
                 invalid => Err(Error::EnumValueNotKnown(
                     "seal::Confidential".to_string(),
